# [inductor] decompose tanh in CPP backend